[Refactor] Change layout to match docker img & create AutoNode lib
