+++ root/balances.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+./bin/hmy balances "$(cat /.val_address)"
