+++ root/cleanse-bls.py
@@ -1,8 +1,14 @@
 #!/usr/bin/env python3
 import argparse
+import time
 from argparse import RawTextHelpFormatter
 
-from utils import *
+from pyhmy import (
+    Typgpy,
+    json_load
+)
+from pyhmy import cli
+import AutoNode
 
 with open("/.val_address", 'r') as f:
     validator_addr = f.read()
@@ -31,7 +37,7 @@ def hard_cleanse():
     """ Only keep BLS keys for current running node.
     """
     # WARNING: Assumption that chain BLS keys are not 0x strings
-    keys_on_chain = get_validator_information(validator_addr, endpoint)['validator']['bls-public-keys']
+    keys_on_chain = AutoNode.get_validator_information(validator_addr, endpoint)['validator']['bls-public-keys']
     for key in keys_on_chain:
         if key not in bls_keys:
             print(f"{Typgpy.WARNING}Removing {key}, key not in node's list of BLS keys: {bls_keys}{Typgpy.ENDC}")
@@ -45,7 +51,7 @@ def shard_cleanse():
     """ Only keep BLS keys on same shard as current running node.
     """
     # WARNING: Assumption that chain BLS keys are not 0x strings
-    keys_on_chain = get_validator_information(validator_addr, endpoint)['validator']['bls-public-keys']
+    keys_on_chain = AutoNode.get_validator_information(validator_addr, endpoint)['validator']['bls-public-keys']
     shard = json_load(cli.single_call(f"hmy utility shard-for-bls {list(bls_keys)[0].replace('0x', '')} "
                                       f"-n {endpoint}"))['shard-id']
     for key in keys_on_chain:
@@ -63,22 +69,22 @@ def reward_cleanse():
     """ Only keep BLS keys that have earned something in the current epoch
     """
     # WARNING: Assumption that chain BLS keys are not 0x strings
-    val_metrics = get_validator_information(validator_addr, endpoint)['metrics']
+    val_metrics = AutoNode.get_validator_information(validator_addr, endpoint)['metrics']
     if val_metrics is None:
         print(f"{Typgpy.WARNING}Can not get current BLS key performance, "
               f"validator ({validator_addr}) is not elected.{Typgpy.ENDC}")
-        if args.yes or input_with_print(f"Wait for election? [Y]/n\n> ") in {'Y', 'y', 'yes', 'Yes'}:
+        if args.yes or AutoNode.input_with_print(f"Wait for election? [Y]/n\n> ") in {'Y', 'y', 'yes', 'Yes'}:
             while val_metrics is None:
                 time.sleep(8)
-                val_metrics = get_validator_information(validator_addr, endpoint)['metrics']
+                val_metrics = AutoNode.get_validator_information(validator_addr, endpoint)['metrics']
         else:
             exit()
-    block_per_epoch = get_metadata(endpoint)['blocks-per-epoch']
+    block_per_epoch = AutoNode.get_metadata(endpoint)['blocks-per-epoch']
     # WARNING: Assumption that epochs are greater than 6 blocks
-    while 0 <= get_latest_header("http://localhost:9500/")['blockNumber'] % block_per_epoch <= 5:
+    while 0 <= AutoNode.get_latest_header("http://localhost:9500/")['blockNumber'] % block_per_epoch <= 5:
         pass
-    bls_metrics = get_validator_information(validator_addr, endpoint)['metrics']['by-bls-key']
-    keys_on_chain = get_validator_information(validator_addr, endpoint)['validator']['bls-public-keys']
+    bls_metrics = AutoNode.get_validator_information(validator_addr, endpoint)['metrics']['by-bls-key']
+    keys_on_chain = AutoNode.get_validator_information(validator_addr, endpoint)['validator']['bls-public-keys']
     for metric in bls_metrics:
         if metric['earned-reward'] == 0:
             key = metric['key']['bls-public-key']
@@ -92,7 +98,6 @@ def reward_cleanse():
 
 if __name__ == "__main__":
     args = parse_args()
-    setup()
     all_val = json_load(cli.single_call(f"hmy --node={endpoint} blockchain validator all"))["result"]
     if validator_addr not in all_val:
         print(f"{Typgpy.FAIL}{validator_addr} is not a validator on {endpoint}.{Typgpy.ENDC}")
