+++ root/docker-setup.sh
@@ -1,0 +1,3 @@
+pip3 install .
+chmod +x ./run.sh ./activate.sh ./attach.sh ./balances.sh ./cleanse-bls.py ./create_validator.sh
+chmod +x ./deactivate.sh ./export.sh ./header.sh ./headers.sh ./info.sh ./version.sh
