+++ root/info.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+./bin/hmy blockchain validator information "$(cat /.val_address)" -n "$(cat /.beacon_endpoint)"
