--- scripts/balances.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-./bin/hmy balances $(cat /.val_address)
