--- scripts/info.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-./bin/hmy blockchain validator information $(cat /.val_address) -n $(cat /.beacon_endpoint)
